% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{data}
\name{macro}
\alias{macro}
\title{Macro}
\format{
Data.frame with 188 obs. of  3 variables

\itemize{
\item{\code{date}}{Date}
\item{\code{gdp}}{real GDP}
\item{\code{lgdp}}{log real GDP}
}
}
\source{
U.S. Bureau of Economic Analysis (GDPC1)
}
\usage{
macro
}
\description{
Quarterly data of US log real GDP from 1970Q1 to 2016Q4
}
\keyword{datasets}
