% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendex.R
\name{plot.tendex}
\alias{plot.tendex}
\title{Methods for mFilter objects}
\usage{
\method{plot}{tendex}(x, reference.grid = TRUE, col = "steelblue", ...)
}
\arguments{
\item{x}{an object of class "trendex"; usually, a result of a call to \code{\link{trendex}}.}

\item{reference.grid}{logical. if true grid lines are drawn.}

\item{col}{color of the graph (see \code{\link[graphics]{plot}}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
"trendex" object
}
\description{
Common methods for all mFilter objects usually created by the mFilter function.
}
