% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendex.R
\name{trendex}
\alias{trendex}
\title{Trend Extraction}
\usage{
trendex(x, ...)
}
\arguments{
\item{x}{a regular time series (ts object).}

\item{...}{Additional arguments to pass to the function \code{\link[Rlibeemd]{ceemdan}}.}
}
\value{
An object of class "trendex".
}
\description{
Trend extraction through the empirical mode
decomposition (EMD) and ensemble EMD (EEMD) from non-stationary time series.
}
